+++ components/vf-tabs/vf-tabs.react.js
@@ -4,12 +4,13 @@
  *
  */
 import { useEffect } from "react";
-import { vfTabs } from "@visual-framework/vf-tabs/vf-tabs";
+import { vfTabs } from "@visual-framework/vf-tabs/vf-tabs.js";
 
 function VfTabs(tabsData) {
   useEffect(() => {
     vfTabs();
   }, []);
+
   return (
     <>
       <div className="vf-tabs">