+++ components/vf-tabs/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.1.11
+
+* Changed: Test cases added for React  [Tracking issue](https://github.com/visual-framework/vf-core/issues/2174)
+
 ### 2.1.10
 
 * Changed: Test cases added for Angular  [Tracking issue](https://github.com/visual-framework/vf-core/issues/1932)